v0.07.11 - preparations to dynamic class instance creation dependently from input string, make sure it is not possible to run any script from frontend, update ratios list with more class names params, rename ratiosCalc folder for better name
