+++ invKeepBackend/ratios/ratiosClasses/DividendYieldRatio.js
